fixing build actions

+++ etc/meta/assemble-release.py
@@ -1,4 +1,7 @@
 # Assemble release folder
+# usage:
+# python3 assemble-release.py
+# python3 assemble-release.py from-dir [dst-dir] [build-from-subdir]
 
 import os
 import sys
@@ -6,27 +9,50 @@ import shutil
 import glob
 import pathlib
 
-target = "ogm_release"
+# folder in which to place release
+target = "ogm-release"
 
-binext = ""
-libext = ".so"
-
+# directory to import from
 _from = "."
+
+# folder relative to _from which contains built binaries
+build_dir = os.path.join(_from, "build")
+
+# replace above with clargs
 if len(sys.argv) >= 2:
     _from = sys.argv[1]
 
-if os.name == 'nt':
+if len(sys.argv) >= 3:
+    target = sys.argv[2]
+
+if len(sys.argv) >= 4:
+    build_dir = os.path.join(_from, sys.argv[3])
+
+if not os.path.isdir(_from) or not os.path.isdir(build_dir):
+    print("'from' directory invalid (check that it is a directory and contains a build/ subfolder)")
+    exit(1)
+
+# folder relative to _from which contains libraries to copy in
+lib_dir = os.path.join(_from, "libs")
+
+# platform-dependent binary/library names
+binext = ""
+libext = ".so"
+if sys.platform == "win32" or os.name == 'nt' or sys.platform == "cygwin":
     binext = ".exe"
     libext = ".dll"
-    
-dlibext = "d" + libext
+elif sys.platform == "darwin":
+    # mac os
+    binext = ""
+    libext = ".dylib"
 
+# remove existing target directory and recreate it
 if os.path.exists(target):
     print ("removing existing " + target)
     shutil.rmtree(target)
+os.makedirs(target)
 
-os.mkdir(target)
-
+# verbose versions (to explain what is happening)
 def copytree(src, dst):
     print("copytree " + src + " -> " + dst)
     return shutil.copytree(src, dst)
@@ -34,6 +60,7 @@ def copy(src, dst):
     print("copy " + src + " -> " + dst)
     return shutil.copy(src, dst)
     
+# sparce copy etc/ folder
 # etc/
 os.mkdir(target + "/etc")
 endings = ["*.png", "*.gif", "*.ico"]
@@ -44,27 +71,30 @@ for ending in endings:
 # demo/
 copytree("demo", target + "/demo")
 
-# binaries
-copy(os.path.join(_from, "ogm" + binext), target)
+# license
+copy(os.path.join(build_dir, "LICENSE"), os.path.join(target, "LICENSE"))
+
+# binaries (including gig)
+copy(os.path.join(build_dir, "ogm" + binext), target)
 os.chmod(target + "/ogm" + binext, 777)
-copy(os.path.join(_from, "ogm-test" + binext), target)
+copy(os.path.join(build_dir, "ogm-test" + binext), target)
 os.chmod(target + "/ogm-test" + binext, 777)
-for file in pathlib.Path(_from).rglob('*' + libext):
-    file = str(file)
-    if os.path.dirname(file) == target:
-        continue
-    print("found " + file)
-    copy(file, target)
-    
-# licenses
-copy("LICENSE", target + "/LICENSE_opengml")
-copy("external/xbr/xbrjs.license", target + "/LICENSE_xbrjs")
-copy("external/pugixml/LICENCE.md", target + "/LICENSE_pugixml")
-copy("external/include/nlohmann/LICENCE.MIT", target + "/LICENSE_nlohmann")
-copy("external/include/rectpack2D/LICENSE.md", target + "/LICENSE_rectpack2d")
-copy("external/include/simpleini/LICENCE.txt", target + "/LICENSE_simpleini")
-copy("external/include/ThreadPool_zlib_license.txt", target + "/LICENCE_ThreadPool")
-copy("external/include/rapidcsv.license", target + "/LICENSE_rapidcsv")
-copy("external/include/base64.license", target + "/LICENSE_base64")
-copy("external/soloud/LICENSE", target + "/LICENSE_soloud")
-copy("external/crossline/LICENSE", target + "/LICENSE_crossline")
+copy(os.path.join(build_dir, "gig" + libext), target)
+
+# copy any desired shared libraries
+if os.path.exists(lib_dir):
+    print(f"searching for additional libraries in {lib_dir}...")
+    any_found = False
+    for file in pathlib.Path(lib_dir).rglob('*' + libext):
+        file = str(file)
+        # skip gig
+        if os.path.basename(file).startswith("gig."):
+            continue
+        # don't copy what's already in the target/ dir
+        if os.path.dirname(file) == target:
+            continue
+        print("found " + file)
+        copy(file, target)
+        any_found = True
+    if not any_found:
+        print("(none found.)")
